requirements.txt
aiohttp==0.21.6
chardet==2.3.0
click==6.6
discord.py==0.11.0
dota2api==1.3.2
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==2.11.1
websockets==3.1
Werkzeug==0.11.11
